requirements.txt
algoliasearch==3.0.0
algoliasearch-django==3.0.0
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==3.2.20
django-cors-headers==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
PyJWT==2.7.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.6
